Organize all extension classes in a similar way in separate files
