+++ ClosedXML/Extensions/DictionaryExtensions.cs
@@ -1,0 +1,20 @@
+// Keep this file CodeMaid organised and cleaned
+using System;
+using System.Collections.Generic;
+using System.Linq;
+
+namespace ClosedXML.Excel
+{
+    internal static class DictionaryExtensions
+    {
+        public static void RemoveAll<TKey, TValue>(this Dictionary<TKey, TValue> dic,
+            Func<TValue, bool> predicate)
+        {
+            var keys = dic.Keys.Where(k => predicate(dic[k])).ToList();
+            foreach (var key in keys)
+            {
+                dic.Remove(key);
+            }
+        }
+    }
+}
