+++ ClosedXML/Extensions/EnumerableExtensions.cs
@@ -1,0 +1,33 @@
+// Keep this file CodeMaid organised and cleaned
+using System;
+using System.Collections.Generic;
+
+namespace ClosedXML.Excel
+{
+    internal static class EnumerableExtensions
+    {
+        public static void ForEach<T>(this IEnumerable<T> source, Action<T> action)
+        {
+            foreach (T item in source)
+                action(item);
+        }
+
+        public static Type GetItemType<T>(this IEnumerable<T> source)
+        {
+            return typeof(T);
+        }
+
+        public static Boolean HasDuplicates<T>(this IEnumerable<T> source)
+        {
+            HashSet<T> distinctItems = new HashSet<T>();
+            foreach (var item in source)
+            {
+                if (!distinctItems.Add(item))
+                {
+                    return true;
+                }
+            }
+            return false;
+        }
+    }
+}
