+++ ClosedXML/Extensions/GuidExtensions.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿// Keep this file CodeMaid organised and cleaned
+using System;
 
 namespace ClosedXML.Extensions
 {
