+++ ClosedXML/Extensions/OpenXmlPartContainerExtensions.cs
@@ -1,3 +1,4 @@
+// Keep this file CodeMaid organised and cleaned
 using DocumentFormat.OpenXml.Packaging;
 using System;
 using System.Linq;
